Make ClassLoaderCache extendable by delegation
We are having a lot of problems in sbt 1.3.0-RC1 with metaspace
utilization. Leaking classlaoders in the ClassLoaderCache is a primary
culprit. I want to be able to modify the ClassLoaderCache implementation
in sbt so that we can manually close it in a number of situations (such
as when the user runs '++' directly or indirectly. To do this, I define
a trait, AbstractClassLoaderCache that specifies the public api of
ClassLoaderCache and then refactor ClassLoaderCache so that the primary
constructor takes an instance of AbstractClassLoaderCache and delegates
to it. Finally, I provide a default implementation that is identical to
the old version.