Dependency: avoid creating tuples.
The Dependency file was a hotspot for allocation of Tuple2 objects.
Given that `resolveNonLocalClass` is a completely private method
(does not seem to be used anywhere else), and that the second value
always flows to a local field assignment, we can do that assignment in
the procedure and avoid creating the tuple.