Optimize source changes calculations
There were a number of intermediate sets generated in the process for
determining which source files had changed since the last compilation
run. We can avoid making intermediate sets by filling four mutable java
sets and then wrapping them in class that implements
Set[VirtualFileRef]. By using the wrapped set, we can avoid extra copies
in most cases. The wrapped scala set is effectively copy on write so as
long as we don't add or remove any elements, we can just delegate to the
underlying set.

The api for Set changed between 2.12 and 2.13 so it was necessary to
provide separate implementations of WrappedSet.