Avoid Files.isDirectory on jar files
Files.isDirectory is relatively expensive and we should avoid it if
possible, which is the case for .jar files. This means that the name of
a classpath directory can't end in .jar. I think this is a reasonable
restriction and I don't think this will ever be the case without the
user explicitly changing the target directory name in sbt or any other
build tool, really.