Fix URI path management
Because I have Artifactory set up as my ~/.sbt/repositories, the
scala-reflect that Zinc's "macros" scripted tests was using was the one
at:

    ~/Library/Caches/Coursier/v1/http/127.0.0.1%3A8081/artifactory/scala-ci-virtual/org/scala-lang/scala-reflect/2.13.2/scala-reflect-2.13.2.jar

but because of how virtual files go through URI.toString, this was
being encoding that file path as:

    ~/Library/Caches/Coursier/v1/http/127.0.0.1%253A8081/artifactory/scala-ci-virtual/org/scala-lang/scala-reflect/2.13.2/scala-reflect-2.13.2.jar

Note the extra "25" (not sure why it's 25, btw).

So, stop URI encoding and use the NIO Path APIs to replace root paths.

Similarly, in LocalJava, use the right URI contructor to minimally
URL-encode and decode the path, adding & removing the needed /-prefix.