Buffer scripted run, too
The reason it's System.out even though the tests use println (aka
Console.out.println) is because (I believe) the classloader is creating
a brand new scala.Console class, which re-initialises it's `object`
state against the (already loaded) java.lang.System.

As such, I'm a little unsure about this patch, as:

1. I'm changing a global variable (System.setOut);
2. and relying on this Console/lazy initialisations setup

... and all this for the benefit of dropping some annoying output in the
test runs.  Suggestions welcome, and a revert wouldn't be a surprise.