compileAllJava
This implements an Analysis-aware `compileAllJava(...)`.

```scala
def compileAllJava(in: Inputs, logger: Logger): CompileResult = ...
```

How this is realized is more complicated than one might expect because we need all the wiring to get AnalysisCallback going to reuse the `Incremental` + `MixedAnalyzingCompiler` harness (This is so the *downstream* subproject can get the incremental goodness).