Fix macro dependency test if !reuseCompilerInstance
With the removal of the CompilerCache/GlobalsCache, no tests are using
the same compiler with multiple Runs (i.e. a "resident" compiler).
Instead they spawn new compilers.

For the macro dependency test this means that the symbol for the macro
code won't point to source files (leading to "class dependencies") but
to classpath entries (leading to "binary dependendencies"), so change
the test accordingly.

Also, finish tearing out all the reference to `reuseCompilerInstance`.