Optimize protobuf writer code
Writing the analysis protobuf file can be very slow when there are many
source files in a project. Part of the problem is that the the protobuf
library was not used as efficiently as it could have been. This commit
goes through and removes all of the bulk put/add operations of the
protobuf builders and instead adds elements one by one. By doing it this
way, we can avoid generating many intermediate java collections. The
upshot was that on my laptop, the time to write the analysis file
dropped from roughly 400ms to roughly 200ms for the project in
https://github.com/eatkins/scala-build-watch-performance when there are
5000 source files.

The reason for this optimization was that I noticed that zinc 1.4.0-M1
was significantly slower in some cases than zinc 1.3.5. I generated a
flamegraph and noticed that the toApis method was very slow in 1.4.0-M1.
It turned out that the putAllInternal method was where most of the time
was spent. When I switched to putting the elements one by one, it
disappeared from the flamegraph. After that, I went through and replaced
all the calls to putAll* and addAll*.