Fix WrappedSet.contains
I noticed from code inspection that the contains method of WrappedSet
was not correctly implemented. It doesn't appear that it is called in
zinc but I figured it would still be good to fix in case it ever is. (My
assumption was that only the iterator method of these Sets would be
called since they just hold the source and binary changes between
analysis runs. Avoiding creating an immutable set when when all we
needed was an iterator over the existing data was the reason for the
WrappedSet class in the first place.)