Use toIterator instead of toStream
Historically Streams have not performed nearly as well as Iterator. I'm
not sure how relevant that is the case for the latest versions of scala
but there were a number of places where it was easy to switch from
Streams to Iterators so this commit does just that. There is very little
expressivity loss in the switch from Iterator to Stream.

From a performance perspective, I have been testing the zinc overhead in
a small project by looping 100 iterations of no-op Test / compile. While
not a scientific benchmark by any means, I pretty consistently found
zinc to run about 10% faster (~63ms average after this change compared
to ~70 ms before).