Cache the jar class definitions
It is slow to compute all of the classes in every jar on the classpath.
We can cache the jar class names between invocations so long as we check
that the timestamp of the jar hasn't changed. My assumption is that the
timestamps of library jars rarely change so the cache entry will rarely
be invalidated. It isn't catastrophic if it is.

To ensure that the caching isn't global, I introduce a class,
Locate.DefinesClass, which provides the same api as Locate.definesClass,
but it can cache the the jar entries. By using a class, we allow the
build tool to control whether or not they want to cache at all.

On my laptop, this reduced the average runtime for a no-op test compile in a
simple project from about 70ms to 60ms. The project had a fairly small
dependency classpath:
org/scala-lang/scala-library/2.13.3/scala-library-2.13.3.jar
org/scalatest/scalatest_2.13/3.0.8/scalatest_2.13-3.0.8.jar
com/lihaoyi/utest_2.13/0.7.4/utest_2.13-0.7.4.jar
org/scalactic/scalactic_2.13/3.0.8/scalactic_2.13-3.0.8.jar
org/scala-lang/scala-reflect/2.13.3/scala-reflect-2.13.3.jar
org/scala-lang/modules/scala-xml_2.13/1.2.0/scala-xml_2.13-1.2.0.jar
org/scala-sbt/test-interface/1.0/test-interface-1.0.jar
org/portable-scala/portable-scala-reflect_2.13/0.1.1/portable-scala-reflect_2.13-0.1.1.jar

so I suspect the difference will be more significant for projects with a
heavier dependency graph.