Only recompile all java sources if any have changed
When any java source is invalidated, we want to recompile the entire
subproject to generate the java class signatures. For this reason, it is
necessary to invalidate _all_ of the java sources when any change is
detected. We don't however want to recompile all of the java sources if
_none_ of them have changed.

Prior to this change, in the sbt build with pipelining off, changing any
scala source in a mixed project would rebuild all of the java sources
which could trigger a cascade of invalidations that could potentially
invalidate the entire project even when none of the java sources had
actually changed. Due to this behavioral change, in
0504f70454fdbe9db2cba4003c710888f6e31e2a it was necessary to change the
iteration count in less-inter-inv-java test from 3 to 2 because the
overcompiling behavior made it only do one cycle when a single source
file was changed because all of the sources were automatically
invalidated. This commit restores the old behavior so the test can be
restored as well. It also adds a test for mixed projects to ensure that
changing a scala file does not automatically trigger recompilation of
java files.