Always create output directories
We should always call Files.createDirectories to ensure the output
directory exists. If the directory already exists and is a directory,
Files.createDirectories is a no-op and is the same amount of io as the
Files.exists check. If the directory exists but is actually a file, then
that is a bug and compilcation is likely to fail so we should throw an
exception.