Reduce io in ClasspathCache
I noticed in a flame graph that we were calling Files.exists and
Files.readAttributes in the same place. It should be unnecessary to call
Files.exists because Files.readAttributes will throw a
NoSuchFileException if it doesn't exist. This would be probably be a bad
change if we were checking files that frequently didn't exist but I
doubt that's the case. We also probably should have had the catch of
NoSuchFileException regardless of whether we did the existence check
because it's always possible albeit unlikely that the file is deleted in
between the existence check and reading the attributes.

I also saw that we could avoid computing the last modified time with
IO.getModifiedTimeOrZero when using jdk > 10 since newer jdks have
millisecond timestamp precision.