Improve incremental compilation on multi module project
Consider a multimodule project with project `core` and project `root`
where `root` depends on `core`, if there are source code changes on
`core` that dont cause invalidations on `root` then the current
`Analysis` is not updated to include the external changes from the
subproject. Subsequent incremental compilations keep considering the
changes despite nothing changing.

The fix is to update the analysis with the external changes even if
there are no invalidations in the current module.

fixes #810