Hold reference to compiler bridge classloader
When sbt is low on memory, sometimes it will evict the classloader that
was constructed for the compiler bridge causing compilation to fail with
a mysterious NoClassDefFoundError. The compiler cache should not evict a
classloader that has any references to it. This commit reworks
AnalyzingCompiler so that we make an explicit reference to the compiler
bridge classloader and pass that loader into the methods that need to
access the classes in the loader instead of passing in the logger from
which we create an ephemeral loader.

After this change, I was unable to reproduce
https://github.com/sbt/zinc/issues/914