Invalidate all sources after transitive step
If there are cyclic dependencies in files, it is possible to get in a
loop where we repeatedly recompile the same files. To avoid that, we can
just invalidate all of the sources if the cycle count exceeds the
IncOptions.transitiveStep parameter. This is a blunt tool but is better
than getting stuck in a loop.