Minimal scala3-compiler-bridge
The `scala3-compiler-bridge` lives in its own module because it shares
almost no sources with the Scala 2 `compiler-bridge`. It is compiled by
Scala 3, using `sbt-dotty`, because it depends on the Scala 3 compiler,
which is compiled in Scala 3. As soon as `sbt-dotty` is merged into
sbt, the `scala3-compiler-bridge` should be able to compiles itself.

This first minimal implementation is a direct translation of
`sbt-bridge` in `https://github.com/lampepfl/dotty`:
- https://github.com/lampepfl/dotty/blob/master/sbt-bridge/src/xsbt/CompilerInterface.java
- https://github.com/lampepfl/dotty/blob/master/sbt-bridge/src/xsbt/CachedCompilerImpl.java

One minor difference is that it implements the new
`xsbti.compile.CompilerInterface2`.