Add support for VirtualFile in Scala 3
This commit add support for `VirtualFiles`, that has recently been added
to Zinc, to the `scala3-compiler-bridge`. It fixes a bug that was
reported in https://github.com/sbt/sbt/pull/6007.

Also it fixes a bug in the bug in the returned `CompileAnalysis` by
populating the `SourceInfos`. The reported warnings were not returned
back in `sbt-dotty` after a succefull compilation.