Improve DualLoader for Scala 3
Change the implementation of `DualLoader` to fix the `LinkageError`
exception that occurs in the Scala 3 compiler when the `xsbti.*` classes
are loaded. The problem is that those classes are loaded by two
different `ClassLoaders`: the sbt loader and the scala instance loader.

In the fixed implementation, the compiler classes are loaded by
the `DualLoader` itself, which extends `URLClassLoader`, so that the
subsequent `xsbti.*` classes can be loaded by the underlying `dual`
loader which is the sbt loader.

See https://github.com/lampepfl/dotty/blob/master/sbt-bridge/src/xsbt/CompilerClassLoader.java
for more details.