Made AnalysisCallback thread-safe again. Re sbt/sbt#6198
For the parallel Hydra Scala compiler to work with sbt and zinc, it's paramount
that the AnalysisCallback is thread-safe. The bulk of the work was done in
sbt/zinc#410, but the changes introduced to support pipelined compilation were
not thread-safe, hence could potentially break compilation via Hydra.

There are three categories of concurrency hazards that are fixed:

* Memory visibility issues (`@volatile` helps with this)
* Compiler/JVM optimizations leading to instructions re-ordering (`@volatile
  prevents this from happening in the sections of interest)
* Race conditions due to concurrent read/write (`AtomicBoolean` CAS heals this)

Mind that instead of `@volatile`, it was perfectly fine to use
`AtomicReference`/`AtomicBoolean`, but I opted for `@volatile` in the places
where CAS were not required. The rationale is to apply KISS.