Drop pruning in compileAllJava
I couldn't find or guess why compileAllJava pruned its previous
analysis, so I removed it and found only "apiinfo/main-discovery"
failed.  That turned out to be because SourceInfos merging was removing
previously found main classes, so I tweaked that.  I reused the logic
for when to prune the analysis anyways (changes in setup).