Force the type of an annotation's typeSymbol beofre checking isStatic
Before 2.13, `isStatic` is implemented by
`isNonBottomSubClass(StaticAnnotationClass)`
which forces the annotation symbol's type.

In 2.13, Java annotations are identified by flags. This check doesn't
force the info, and the flags are missing if the info is still a
`ClassfileLoader`.

This leads to spurious API changes (annotation goes missing) if the
depending if the info is already forced or not.

A fix for this will be in 2.13.7, but we should still work around it
in Zinc to make sure zinc works correctly on 2.13.0-6.