remove no-longer-needed IntelliJ workaround
the workaround suggested in build.sbt for importing the build
into IntelliJ doesn't work anymore with the current versions
of things (not sure when it broke or why)

this is my attempt to do better and it seems to work. I'm not
sure yet if it's perfect, but it's a clear improvement on the
status quo. we can always adjust further later, as needed