Overwrite Problem#rendered with None when source mapping occurs.
This commit is to improve Twirl's error message with Scala3.
https://github.com/playframework/twirl/issues/498

This pull request makes sbt ignore `Problem#rendered` passed from the compiler  if source mapping is actually performed.
Sbt will display the error message based on the `Position` transformed by source mapping.
When the source mapping is not performed, the message in `Problem#rendered` is displayed as before.